Enum relay_protocol::Val

pub enum Val<'a> {
    Bool(bool),
    I64(i64),
    U64(u64),
    F64(f64),
    String(&'a str),
    Uuid(Uuid),
    Array(Arr<'a>),
    Object(Obj<'a>),
}

Borrowed version of Value.

Variants

Bool(bool)

A boolean value.

I64(i64)

A signed integer value.

U64(u64)

An unsigned integer value.

F64(f64)

A floating point value.

String(&'a str)

A string value.

Uuid(Uuid)

A UUID.

Array(Arr<'a>)

An array of annotated values.

Object(Obj<'a>)

A mapping of strings to annotated values.

Implementations

impl<'a> Val<'a>

pub fn as_bool(&self) -> Option<bool>

Returns the value if it is a boolean, otherwise None.

pub fn as_i64(&self) -> Option<i64>

Represents the value as f64 if possible. Returns None otherwise.

pub fn as_u64(&self) -> Option<u64>

Represents the value as f64 if possible. Returns None otherwise.

pub fn as_f64(&self) -> Option<f64>

Represents the value as f64 if possible. Returns None otherwise.

pub fn as_str(&self) -> Option<&'a str>

Returns the string if this value is a string, otherwise None.

pub fn as_uuid(&self) -> Option<Uuid>

Returns the UUID if this value is a UUID, otherwise None.