Enum relay_protocol::Value

pub enum Value {
    Bool(bool),
    I64(i64),
    U64(u64),
    F64(f64),
    String(String),
    Array(Array<Value>),
    Object(Object<Value>),
}

Represents a boxed value.

Variants

Bool(bool)

A boolean value.

I64(i64)

A signed integer value.

U64(u64)

An unsigned integer value.

F64(f64)

A floating point value.

String(String)

A string value.

Array(Array<Value>)

An array of annotated values.

Object(Object<Value>)

A mapping of strings to annotated values.

Implementations

impl Value

pub fn describe(&self) -> ValueDescription<'_>

Returns a formattable that gives a helper description of the value.

pub fn as_str(&self) -> Option<&str>

Returns the string if this value is a string, otherwise None.