Struct relay_protocol::Annotated

pub struct Annotated<T>(pub Option<T>, pub Meta);

Wrapper for data fields with optional meta data.

Tuple Fields
0: Option<T>1: Meta
Implementations

impl<T> Annotated<T>

pub fn new(value: T) -> Self

Creates a new annotated value without meta data.

pub fn empty() -> Self

Creates an empty annotated value without meta data.

pub fn from_error<E>(err: E, value: Option<Value>) -> Selfwhere E: Into<Error>,

Creates an empty annotated value with error attached.

pub fn value(&self) -> Option<&T>

Returns a reference to the value.

Returns None if this value is not initialized, missing, or invalid.

pub fn value_mut(&mut self) -> &mut Option<T>

Returns a mutable reference to the value.

Returns None if this value is not initialized, missing, or invalid.

pub fn set_value(&mut self, value: Option<T>)

Replaces the stored value.

pub fn meta(&self) -> &Meta

Returns a reference to the meta data attached to this value.

pub fn meta_mut(&mut self) -> &mut Meta

Returns a mutable reference to the meta data attached to this value.

pub fn into_value(self) -> Option<T>

Returns the stored value and drops all meta data.

pub fn and_then<F, U, R>(self, f: F) -> Annotated<U>where F: FnOnce(T) -> R, R: Into<Annotated<U>>,

Calls f with the stored value if available and merges meta data into the result.

pub fn map_value<U, F>(self, f: F) -> Annotated<U>where F: FnOnce(T) -> U,

Maps an Annotated<T> to an Annotated<U> and keeps the original meta data.

pub fn get_or_insert_with<F>(&mut self, f: F) -> &mut Twhere F: FnOnce() -> T,

Inserts a value computed from f into the value if it is None, then returns a mutable reference to the contained value.

impl<T> Annotated<T>where T: AsRef<str>,

pub fn as_str(&self) -> Option<&str>

Returns a reference to the string value if set.

impl Annotated<Value>

pub fn as_str(&self) -> Option<&str>

Returns a reference to the string value if this value is a string and it is set.

impl<T> Annotated<T>where T: Empty,

pub fn skip_serialization(&self, behavior: SkipSerialization) -> bool

Returns whether this value should be skipped during serialization.

An Annotated<T> is always serialized if it has meta data. Otherwise, serialization depends on the behavior. For SkipSerialization::Empty, the Empty trait is used to determine emptiness of the contained value and defaults to false for no value.

impl<T> Annotated<T>where T: FromValue,

pub fn deserialize_with_meta<'de, D: Deserializer<'de>>( deserializer: D ) -> Result<Self, D::Error>

Deserializes an annotated from a deserializer

pub fn from_json(s: &str) -> Result<Self, Error>

Deserializes an annotated from a JSON string.

pub fn from_json_bytes(b: &[u8]) -> Result<Self, Error>

Deserializes an annotated from JSON bytes.

impl<T> Annotated<T>where T: IntoValue,

pub fn serialize_with_meta<S: Serializer>( &self, serializer: S ) -> Result<S::Ok, S::Error>

Serializes an annotated value into a serializer.

pub fn to_json(&self) -> Result<String, Error>

Serializes an annotated value into a JSON string.

pub fn to_json_pretty(&self) -> Result<String, Error>

Serializes an annotated value into a pretty JSON string.

pub fn payload_to_json(&self) -> Result<String, Error>

Serializes an annotated value into a JSON string.

pub fn payload_to_json_pretty(&self) -> Result<String, Error>

Serializes an annotated value into a pretty JSON string.