relay_protocol

Trait FromObjectRef

pub trait FromObjectRef: FromValue {
    // Required method
    fn from_object_ref(value: &mut Object<Value>) -> Self
       where Self: Sized;
}

Implemented for all meta structures which can be created from an object.

Only meta structures which implement FromObjectRef can be flattened.

Required Methods

fn from_object_ref(value: &mut Object<Value>) -> Self
where Self: Sized,

Creates a meta structure from key value pairs.

The implementation is supposed remove used fields from the passed value.