Module relay_protocol::condition

Types to specify conditions on data.

The root type is RuleCondition.

Structs

AndCondition

Combines multiple conditions using logical AND.

EqCondOptions

Options for EqCondition.

EqCondition

A condition that compares values for equality.

GlobCondition

A condition that uses glob matching.

GtCondition

A condition that applies >.

GteCondition

A condition that applies >=.

LtCondition

A condition that applies <.

LteCondition

A condition that applies <=.

NotCondition

Applies logical NOT to a condition.

OrCondition

Combines multiple conditions using logical OR.

Enums

RuleCondition

A condition that can be evaluated on structured data.

Traits

IntoStrings

A type that can be converted to a list of strings.